add - changed the Id type to be int - VC

+++ api/backend/Models/CartItem.cs
@@ -6,8 +6,8 @@ namespace backend.Models
     public class CartItem
     {
         [Key]
-        public long Id { get; set; }
-        public long PaintingId { get; set; }
+        public int Id { get; set; }
+        public int PaintingId { get; set; }
         public int Quantity { get; set; }
         public virtual Painting Painting { get; set; }
     }
